Split encode module into encodinghelpers, encodemsg, encode
